feat: extract chronological order to own file
